Update windows M1
We also update a correction in the code for the mac os GUI.

+++ Xcode/lispegui/LispE/Linenumber.h
@@ -24,7 +24,7 @@
     NSMutableArray* ypositions;
 }
 
--(bool)findbreakline:(int)ligne;
+-(bool)findbreakline:(long)ligne;
 -(void)ajoutey:(NSInteger)i;
 -(void)nettoiey;
 -(void)ajoutel:(NSInteger)i;
